Refactor package names to use consistent lowercase naming
Updated all package names in the backend to follow lowercase naming conventions for consistency and readability. Adjusted references accordingly to support the changes.
