--- app/backend/CommandManager/CommandManager.go
@@ -1,4 +1,0 @@
-package CommandManager
-
-type CommandManager struct {
-}
