--- app/backend/Component/Association.go
@@ -1,4 +1,0 @@
-package Component
-
-type Association struct {
-}
