--- app/backend/Component/Component.go
@@ -1,4 +1,0 @@
-package Component
-
-type Component interface {
-}
