--- app/backend/Component/Gadget.go
@@ -1,4 +1,0 @@
-package Component
-
-type Gadget struct {
-}
