--- app/backend/UMLDiagram/UMLDiagram.go
@@ -1,4 +1,0 @@
-package UMLDiagram
-
-type UMLDiagram struct {
-}
