--- app/backend/UMLProject/UMLProject.go
@@ -1,4 +1,0 @@
-package UMLProject
-
-type UMLProject struct {
-}
