--- app/backend/Utils/ConnectionError.go
@@ -1,13 +1,0 @@
-package Utils
-
-type ConnectionError struct {
-	msg string
-}
-
-func (e *ConnectionError) Error() string {
-	return e.msg
-}
-
-func NewConnectionError(msg string) *ConnectionError {
-	return &ConnectionError{msg: msg}
-}
