--- app/backend/Utils/DUError.go
@@ -1,5 +1,0 @@
-package Utils
-
-type DUError interface {
-	Error() string
-}
