--- app/backend/Utils/FileIOError.go
@@ -1,13 +1,0 @@
-package Utils
-
-type FileIOError struct {
-	msg string
-}
-
-func (e *FileIOError) Error() string {
-	return e.msg
-}
-
-func NewFileIOError(msg string) *FileIOError {
-	return &FileIOError{msg: msg}
-}
