--- app/backend/Utils/MemoryFullError.go
@@ -1,13 +1,0 @@
-package Utils
-
-type MemoryFullError struct {
-	msg string
-}
-
-func (e *MemoryFullError) Error() string {
-	return e.msg
-}
-
-func NewMemoryFullError(msg string) *MemoryFullError {
-	return &MemoryFullError{msg: msg}
-}
