--- app/backend/Utils/SendError.go
@@ -1,13 +1,0 @@
-package Utils
-
-type SendError struct {
-	msg string
-}
-
-func (e *SendError) Error() string {
-	return e.msg
-}
-
-func NewSendError(msg string) *SendError {
-	return &SendError{msg: msg}
-}
