+++ app/backend/command/CommandManager.go
@@ -1,0 +1,4 @@
+package command
+
+type CommandManager struct {
+}
