+++ app/backend/component/Association.go
@@ -1,0 +1,4 @@
+package component
+
+type Association struct {
+}
