+++ app/backend/component/Component.go
@@ -1,0 +1,4 @@
+package component
+
+type Component interface {
+}
