+++ app/backend/component/Gadget.go
@@ -1,0 +1,4 @@
+package component
+
+type Gadget struct {
+}
