+++ app/backend/umldiagram/UMLDiagram.go
@@ -1,0 +1,4 @@
+package umldiagram
+
+type UMLDiagram struct {
+}
