+++ app/backend/umlproject/UMLProject.go
@@ -1,0 +1,4 @@
+package umlproject
+
+type UMLProject struct {
+}
