+++ app/backend/utils/ConnectionError.go
@@ -1,0 +1,13 @@
+package utils
+
+type ConnectionError struct {
+	msg string
+}
+
+func (e *ConnectionError) Error() string {
+	return e.msg
+}
+
+func NewConnectionError(msg string) *ConnectionError {
+	return &ConnectionError{msg: msg}
+}
