+++ app/backend/utils/DUError.go
@@ -1,0 +1,5 @@
+package utils
+
+type DUError interface {
+	Error() string
+}
