+++ app/backend/utils/InvalidArgumentError.go
@@ -1,4 +1,4 @@
-package Utils
+package utils
 
 type InvalidArgumentError struct {
 	msg string
